Upload of early dev Mapbox instance
This is a very early development instance of the Mapbox Spatial Viewer.  Contained within an HTML file (for now).  Will need to be chopped up and cleaned for the backend of the platform.  

This is pulling the GeoJSONs from UBC Spatial Data github.  Still a lot of data to add, as well as UI additions and tweaks.